parseQueryStringを共通ユーティリティへ分離し、特殊キー対策とログ出力理由を明文化 #2493
- core/src/url_util.mts を新設し parseQueryString を移動(plugin_system_url と plugin_httpserver の双方から参照)
- console.errorへの変更理由(#64061のIPC破損対策)をコメント化
- __proto__が値として取得できることとprototype汚染がないことをHTTPサーバテストで検証
